requirements.txt
channels_redis
djangorestframework
pillow
django-environ
django-cors-headers
djangorestframework-simplejwt